Fix the handling of quoted extra args
So far it was impossible to pass args containing whitespace, and thus
requiring quoting, to the %gobuild macro as the quoting was used
inconsistently. This caused the quoted arguments to be split up at the
whitespace at some point, making it impossible to e.g. pass extra
linker flags to set variable values via `-X`. This is e.g. required
when building stuff using golang-github-prometheus-common, which
expects the version information to be set via
`"-X
%{import_path}/vendor/github.com/prometheus/common/version.Version=%{version}"`.